+++ gluten-core/src/main/scala/org/apache/spark/util/SparkDirectoryUtil.scala
@@ -78,12 +78,9 @@ object SparkDirectoryUtil extends Logging {
       INSTANCE = new SparkDirectoryUtil(roots)
       return
     }
-    if (INSTANCE.roots.toSet != roots.toSet) {
-      logWarning(
-        s"Reinitialize SparkDirectoryUtil with different root dirs: old: ${INSTANCE.ROOTS
-            .mkString("Array(", ", ", ")")}, new: ${roots.mkString("Array(", ", ", ")")}"
-      )
-    }
+    throw new UnsupportedOperationException(
+      s"SparkDirectoryUtil was already initialized with root dirs: ${INSTANCE.ROOTS
+          .mkString("Array(", ", ", ")")}")
   }
 
   def get(): SparkDirectoryUtil = synchronized {
